EONEXT-137 - Added custom terms for the advanced search page. Fix for development config.
