+++ src/core/utils/config.tsx
@@ -1,7 +1,6 @@
 import { RootState, useSelector } from "../store";
 import { addConfigEntries } from "../config.slice";
 import withSuffix from "./withSuffix";
-import serviceArgs from "../storybook/serviceUrlArgs";
 
 function config(key: string): string;
 
@@ -66,48 +65,3 @@ export const withConfig = <T extends object>(
 };
 
 export default {};
-
-if (typeof window === "object" && serviceArgs.developmentOptions === "true") {
-  const extendedCovers = "cover.detail";
-  const extendedFields = {
-    description: {
-      Type: {
-        data: ["marc:001.a", "marc:001.c"],
-        insert: "prepend"
-      },
-      Emnetal: {
-        data: ["marc:088.a"],
-        insert: "replace",
-        url: "/search?q=${tag}"
-      },
-      "Skøn-/faglitteratur": {
-        hidden: true
-      },
-      Emneord: {
-        data: ["marc:631.a"],
-        insert:"prepend",
-        url:"/search?q=${tag}"
-      }
-    },
-    detail: {
-      Type: {
-        data: ["marc:001.a", "marc:001.c"],
-        insert: "prepend"
-      },
-      Stemmer: {
-        data: ["marc:509.a"]
-      },
-      "Stemmer forkortet": {
-        data: ["marc:509.b"]
-      },
-      Indhold: {
-        data: ["marc:795.a","marc:530.a"],
-        type: "list",
-        insert:"fallback"
-      }
-    }
-  };
-
-  document.body.setAttribute("data-eonext-ext-covers", extendedCovers);
-  document.body.setAttribute("data-eonext-ext-fields", JSON.stringify(extendedFields));
-}
